modified the elementwise_op_broadcast and elementwise_op_impl for xpu2
