+++ paddle/fluid/operators/elementwise/elementwise_op_impl.cu.h
@@ -25,6 +25,10 @@ limitations under the License. */
 #define ELEMENTWISE_BLOCK_SIZE 512
 #endif
 
+#ifdef PADDLE_WITH_XPU
+#define ELEMENTWISE_BLOCK_SIZE 64
+#endif
+
 namespace paddle {
 namespace operators {
 
@@ -72,6 +76,148 @@ int GetVectorizedSizeForTensors(
   return vec_size;
 }
 
+struct DimensionsTransform {
+  using DimVector = std::vector<int64_t>;
+  typedef void (*MergeFunctor)(bool &, std::vector<DimVector> &, DimVector &,
+                               int, int);
+  int64_t dim_size;
+  DimVector out_dims;
+  std::vector<DimVector> in_dims;
+
+ private:
+  // To compensate the lackage of input_tensors` dimension with input variable
+  // 'axis'
+  void InputDimensionsExtend(int N, int axis) {
+    for (auto &in_dim : in_dims) {
+      int64_t in_idx = 0;
+      if (in_dim.size() < dim_size) {
+        DimVector tmp_dim(dim_size, 1);
+        do {
+          if (in_dim[in_idx] == out_dims[axis] || in_dim[in_idx] == 1) {
+            tmp_dim[axis] = in_dim[in_idx];
+            in_idx++;
+            axis++;
+          } else {
+            PADDLE_THROW(platform::errors::InvalidArgument(
+                "The %d-th dimension of input tensor is expected to be equal "
+                "with the %d-th dimension of output tensor %d or 1, but "
+                "recieved %d.",
+                in_idx + 1, axis + 1, out_dims[axis], in_dim[in_idx]));
+          }
+        } while (in_idx < in_dim.size());
+        in_dim.resize(dim_size);
+        std::copy(tmp_dim.begin(), tmp_dim.end(), in_dim.begin());
+      } else {
+        do {
+          if (in_dim[in_idx] == out_dims[in_idx] || in_dim[in_idx] == 1) {
+            in_idx++;
+          } else {
+            PADDLE_THROW(platform::errors::InvalidArgument(
+                "The %d-th dimension of input tensor is expected to be equal "
+                "with the %d-th dimension of output tensor %d or 1, but "
+                "recieved %d.",
+                in_idx + 1, in_idx + 1, out_dims[in_idx], in_dim[in_idx]));
+          }
+        } while (in_idx < dim_size);
+      }
+      std::reverse(in_dim.begin(), in_dim.end());
+    }
+    std::reverse(out_dims.begin(), out_dims.end());
+  }
+
+  template <typename MergeFunctor>
+  __inline__ void MergeDimensions(MergeFunctor merge_func, int N) {
+    auto VectorReorganise = [](DimVector *vec, int l_idx, int m_idx) {
+      (*vec)[m_idx - 1] =
+          std::accumulate(vec->begin() + l_idx, vec->begin() + m_idx, 1,
+                          std::multiplies<int64_t>());
+      vec->erase(vec->begin() + l_idx, vec->begin() + m_idx - 1);
+    };
+
+    int64_t i = 0;
+    while (i < dim_size) {
+      int cnt = 0;
+      int low_idx = i;
+      bool equal = true;
+      do {
+        merge_func(equal, in_dims, out_dims, i, N);
+        if (equal) {
+          i++;
+          cnt++;
+        } else {
+          break;
+        }
+      } while (i < dim_size);
+
+      if (cnt > 1) {
+        for (auto &in_dim : in_dims) {
+          VectorReorganise(&in_dim, low_idx, i);
+        }
+        VectorReorganise(&out_dims, low_idx, i);
+        dim_size -= --cnt;
+        i -= cnt;
+      } else if (cnt < 1) {
+        i++;
+      }
+    }
+  }
+
+ public:
+  DimensionsTransform() {}
+
+  explicit DimensionsTransform(
+      const std::vector<const framework::Tensor *> &ins,
+      const framework::DDim &dims, int axis) {
+    const int N = ins.size();
+    dim_size = dims.size();
+    out_dims = framework::vectorize<int64_t>(dims);
+    in_dims.resize(N);
+    for (int j = 0; j < N; ++j) {
+      in_dims[j] = framework::vectorize<int64_t>(ins[j]->dims());
+    }
+    InputDimensionsExtend(N, axis);
+
+    auto merge_sequential_dims = [](bool &equal,
+                                    std::vector<DimVector> &in_dims,
+                                    DimVector &out, int i, int num) {
+      for (int j = 1; j < num; ++j) {
+        equal = (in_dims[0][i] == in_dims[j][i]) ? true : false;
+      }
+    };
+    auto merge_sequential_one_dims = [](bool &equal,
+                                        std::vector<DimVector> &in_dims,
+                                        DimVector &out, int i, int num) {
+      equal = in_dims[0][i] == 1;
+      if (equal) {
+        for (int j = 1; j < num; ++j) {
+          equal = in_dims[j][i] == out[i];
+        }
+      }
+    };
+    // To Merge the dimensions of input_tensors while the consequtive
+    // equal-dimensions appears.
+    MergeFunctor merge_ptr = merge_sequential_dims;
+    MergeDimensions<MergeFunctor>(merge_ptr, N);
+
+    int min_idx = 0;
+    int min_val = std::accumulate(in_dims[0].begin(), in_dims[0].end(), 1,
+                                  std::multiplies<int64_t>());
+    for (int j = 1; j < N; ++j) {
+      int temp = std::accumulate(in_dims[j].begin(), in_dims[j].end(), 1,
+                                 std::multiplies<int64_t>());
+      min_val = min_val > temp ? temp : min_val;
+      min_idx = min_val == temp ? j : min_idx;
+    }
+    std::swap(in_dims[0], in_dims[min_idx]);
+
+    // To Merge the dimension of input_tensors while the consequtive
+    // 1-value-dimensions appears.
+    merge_ptr = merge_sequential_one_dims;
+    MergeDimensions<MergeFunctor>(merge_ptr, N);
+    std::swap(in_dims[min_idx], in_dims[0]);
+  }
+};
+
 template <typename InT, typename OutT, int VecSize, typename Functor, int Arity,
           bool CallElementwiseAny = false>
 struct ElementwisePrimitiveCaller {
@@ -115,64 +261,160 @@ struct ElementwisePrimitiveCaller<InT, OutT, VecSize, Functor, 3, false> {
   }
 };
 
-template <typename InT, typename OutT, typename Functor, int Arity, int VecSize,
-          bool IsBoundary>
-__device__ void DealSegment(
-    const framework::Array<const InT *__restrict__, Arity> &in, OutT *out,
-    int num, Functor func) {
+template <typename T, int VecSize, int Rank, bool IsBoundary = false>
+__device__ __forceinline__ void LoadData(
+    T *dst, const T *__restrict__ src, uint32_t block_offset,
+    const kps::details::BroadcastConfig<Rank> &config, int numel, int num,
+    bool need_broadcast) {
+  // numel : whole num of output
+  // num: how many data will be deal with in this time
+  if (need_broadcast) {
+    kps::ReadDataBc<T, VecSize, 1, 1, Rank, IsBoundary>(dst, src, block_offset,
+                                                        config, numel);
+  } else {
+    kps::ReadData<T, VecSize, 1, 1, IsBoundary>(dst, src + block_offset, num);
+  }
+}
+
+template <typename InT, typename OutT, typename Functor, int VecSize, int Arity,
+          int Rank, bool IsBoundary = false>
+__device__ void ElementwiseKernelImpl(
+    const framework::Array<const InT *__restrict__, Arity> &ins, OutT *out,
+    const framework::Array<bool, Arity> &use_broadcast, uint32_t numel,
+    const framework::Array<kps::details::BroadcastConfig<Rank>, Arity> &configs,
+    int block_offset, Functor func) {
   InT args[Arity][VecSize];
   OutT result[VecSize];
 
-  int data_offset = VecSize * blockIdx.x * blockDim.x;
-
+  int num = numel - block_offset;
 #pragma unroll
   for (int i = 0; i < Arity; i++) {
     kps::Init<InT, VecSize>(args[i], static_cast<InT>(1.0f));
-    kps::ReadData<InT, VecSize, 1, 1, IsBoundary>(args[i], in[i] + data_offset,
-                                                  num);
+    LoadData<InT, VecSize, Rank, IsBoundary>(args[i], ins[i], block_offset,
+                                             configs[i], numel, num,
+                                             use_broadcast[i]);
   }
 
   const bool kCallElementwiseAny =
       platform::FunctionTraits<Functor>::has_pointer_args;
   ElementwisePrimitiveCaller<InT, OutT, VecSize, Functor, Arity,
                              kCallElementwiseAny>()(func, args, result);
-  kps::WriteData<OutT, VecSize, 1, 1, IsBoundary>(out + data_offset, result,
+
+  kps::WriteData<OutT, VecSize, 1, 1, IsBoundary>(out + block_offset, result,
                                                   num);
 }
 
-template <typename InT, typename OutT, typename Functor, int Arity, int VecSize>
-__global__ void ElementVectorizeKernel(
-    framework::Array<const InT *__restrict__, Arity> ins, OutT *out, int size,
-    Functor func) {
-  int data_offset = VecSize * blockIdx.x * blockDim.x;
-  int num = size - data_offset;
-  // the num this time have to deal with
-  if (VecSize * blockDim.x > num) {  // reminder segment
-    DealSegment<InT, OutT, Functor, Arity, VecSize, true>(ins, out, num, func);
-  } else {  // complete segment
-    DealSegment<InT, OutT, Functor, Arity, VecSize, false>(ins, out, num, func);
+template <typename InT, typename OutT, typename Functor, int VecSize, int Arity,
+          int Rank>
+__global__ void ElementwiseKernel(
+    framework::Array<const InT *__restrict__, Arity> ins, OutT *out,
+    framework::Array<bool, Arity> use_broadcast, uint32_t numel,
+    framework::Array<kps::details::BroadcastConfig<Rank>, Arity> configs,
+    int main_tid, Functor func) {
+#ifdef PADDLE_WITH_XPU
+  int data_offset = VecSize * cluster_id() * core_num();
+  int stride = cluster_num() * ncores * VecSize;
+  for (int offset = data_offset; offset < numel; offset += stride) {
+    if (offset + ncores * VecSize < numel) {
+      ElementwiseKernelImpl<InT, OutT, Functor, VecSize, Arity, Rank, false>(
+          ins, out, use_broadcast, numel, configs, offset, func);
+    } else {
+      ElementwiseKernelImpl<InT, OutT, Functor, VecSize, Arity, Rank, true>(
+          ins, out, use_broadcast, numel, configs, offset, func);
+    }
   }
+#else
+  int block_offset = VecSize * blockIdx.x * blockDim.x;
+  if (blockIdx.x < main_tid) {
+    // data offset of this block
+    ElementwiseKernelImpl<InT, OutT, Functor, VecSize, Arity, Rank, false>(
+        ins, out, use_broadcast, numel, configs, block_offset, func);
+  } else {  // reminder
+    ElementwiseKernelImpl<InT, OutT, Functor, VecSize, Arity, Rank, true>(
+        ins, out, use_broadcast, numel, configs, block_offset, func);
+  }
+#endif
 }
 
-template <typename InT, typename OutT, typename Functor, int Arity, int VecSize>
-void ElementwiseCudaKernel(const platform::CUDADeviceContext &ctx,
-                           const std::vector<const framework::Tensor *> &ins,
-                           std::vector<framework::Tensor *> *outs,
-                           Functor func) {
-  auto numel = ins[0]->numel();
-  int block_size = GetThreadsConfig(ctx, numel, VecSize);
-  int grid_size =
-      ((numel + VecSize - 1) / VecSize + block_size - 1) / block_size;
+template <typename InT, typename OutT, typename Functor, int VecSize, int Arity,
+          int Rank>
+void LaunchKernel(const platform::CUDADeviceContext &ctx,
+                  const std::vector<const framework::Tensor *> &ins,
+                  framework::Tensor *out, Functor func,
+                  DimensionsTransform merge_dims,
+                  framework::Array<bool, Arity> use_broadcast) {
+  int numel = out->numel();
+  const int threads = GetThreadsConfig(ctx, numel, VecSize);
+  int blocks = ((numel + VecSize - 1) / VecSize + threads - 1) / threads;
 
+  int main_tid = numel / (VecSize * threads);
   auto stream = ctx.stream();
-  OutT *out_data = (*outs)[0]->data<OutT>();
+  OutT *out_data = out->data<OutT>();
+
+  framework::Array<kps::details::BroadcastConfig<Rank>, Arity> configs;
   framework::Array<const InT *__restrict__, Arity> ins_data;
+
   for (int i = 0; i < Arity; i++) {
     ins_data[i] = ins[i]->data<InT>();
+    if (use_broadcast[i]) {
+      // get the broadcast config,
+      // if data shape is[m, n], then you should set data_dim = {n, m}
+      // eg: out's shape [3, 45, 1]. then out_dims = {1, 45, 3}
+      configs[i] = kps::details::BroadcastConfig<Rank>(
+          merge_dims.out_dims, merge_dims.in_dims[i], merge_dims.dim_size);
+    }
   }
-  ElementVectorizeKernel<InT, OutT, Functor, Arity,
-                         VecSize><<<grid_size, block_size, 0, stream>>>(
-      ins_data, out_data, numel, func);
+
+#ifdef PADDLE_WITH_XPU
+  blocks = 8;
+  ElementwiseKernel<InT, OutT, Functor, VecSize, Arity,
+                    Rank><<<blocks, threads, stream>>>(
+      ins_data, out_data, use_broadcast, numel, configs, main_tid, func);
+#else
+  ElementwiseKernel<InT, OutT, Functor, VecSize, Arity,
+                    Rank><<<blocks, threads, 0, stream>>>(
+      ins_data, out_data, use_broadcast, numel, configs, main_tid, func);
+#endif
+}
+
+template <typename InT, typename OutT, typename Functor, int Arity,
+          int Rank = 1>
+void LaunchElementwiseKernel(const platform::CUDADeviceContext &ctx,
+                             const std::vector<const framework::Tensor *> &ins,
+                             framework::Tensor *out, Functor func,
+                             DimensionsTransform merge_dims,
+                             framework::Array<bool, Arity> use_broadcast) {
+  int in_vec_size = 4;
+  for (auto *in : ins) {
+    auto temp_size = platform::GetVectorizedSize<InT>(in->data<InT>());
+    in_vec_size = in->dims() == out->dims() ? std::min(temp_size, in_vec_size)
+                                            : in_vec_size;
+  }
+  int out_vec_size = platform::GetVectorizedSize<OutT>(out->data<OutT>());
+  int vec_size = std::min(out_vec_size, in_vec_size);
+
+  switch (vec_size) {
+    case 4: {
+      LaunchKernel<InT, OutT, Functor, 4, Arity, Rank>(
+          ctx, ins, out, func, merge_dims, use_broadcast);
+      break;
+    }
+    case 2: {
+      LaunchKernel<InT, OutT, Functor, 2, Arity, Rank>(
+          ctx, ins, out, func, merge_dims, use_broadcast);
+      break;
+    }
+    case 1: {
+      LaunchKernel<InT, OutT, Functor, 1, Arity, Rank>(
+          ctx, ins, out, func, merge_dims, use_broadcast);
+      break;
+    }
+    default: {
+      PADDLE_THROW(platform::errors::Unimplemented(
+          "Unsupported vectorized size: %d !", vec_size));
+      break;
+    }
+  }
 }
 
 template <ElementwiseType ET, typename InT, typename OutT, typename Functor>
@@ -189,28 +431,15 @@ void LaunchSameDimsElementwiseCudaKernel(
                         "arity of functor. But recieved: the number of inputs "
                         "is %d, the arity of functor is %d.",
                         ins.size(), kArity));
-
-  // calculate the max vec_size for all ins and outs
-  int vec_size = GetVectorizedSizeForTensors<InT, OutT>(ins, *outs);
-  switch (vec_size) {
-    case 4:
-      ElementwiseCudaKernel<InT, OutT, Functor, kArity, 4>(ctx, ins, outs,
-                                                           func);
-      break;
-    case 2:
-      ElementwiseCudaKernel<InT, OutT, Functor, kArity, 2>(ctx, ins, outs,
-                                                           func);
-      break;
-    case 1:
-      ElementwiseCudaKernel<InT, OutT, Functor, kArity, 1>(ctx, ins, outs,
-                                                           func);
-      break;
-    default: {
-      PADDLE_THROW(platform::errors::Unimplemented(
-          "Unsupported vectorized size: %d !", vec_size));
-      break;
-    }
+  framework::Tensor *out = (*outs)[0];
+  DimensionsTransform merge_dims;
+  framework::Array<bool, kArity> use_broadcast;
+  for (int i = 0; i < kArity; i++) {
+    use_broadcast[i] = false;
   }
+
+  LaunchElementwiseKernel<InT, OutT, Functor, kArity>(
+      ctx, ins, out, func, merge_dims, use_broadcast);
 }
 
 }  // namespace operators
